rough fix needs work
