+++ diesel_cli/src/config.rs
@@ -1,11 +1,14 @@
 use clap::ArgMatches;
-use serde::Deserialize;
-use std::env;
+use serde::de::{self, MapAccess, Visitor};
+use serde::{Deserialize, Deserializer};
+use serde_regex::Serde as RegexWrapper;
 use std::error::Error;
 use std::fs;
 use std::path::{Path, PathBuf};
+use std::{env, fmt};
 
 use super::find_project_root;
+use crate::infer_schema_internals::TableName;
 use crate::print_schema;
 use crate::print_schema::ColumnSorting;
 
@@ -56,7 +59,7 @@ pub struct PrintSchema {
     #[serde(default)]
     pub with_docs: print_schema::DocConfig,
     #[serde(default)]
-    pub filter: print_schema::Filtering,
+    pub filter: Filtering,
     #[serde(default)]
     pub column_sorting: ColumnSorting,
     #[serde(default)]
@@ -127,5 +130,87 @@ impl MigrationsDirectory {
         if self.dir.is_relative() {
             self.dir = base.join(&self.dir);
         }
+    }
+}
+
+type Regex = RegexWrapper<::regex::Regex>;
+
+pub enum Filtering {
+    OnlyTables(Vec<Regex>),
+    ExceptTables(Vec<Regex>),
+    None,
+}
+
+#[allow(clippy::derivable_impls)] // that's not supported on rust 1.65
+impl Default for Filtering {
+    fn default() -> Self {
+        Filtering::None
+    }
+}
+
+impl Filtering {
+    pub fn should_ignore_table(&self, name: &TableName) -> bool {
+        use self::Filtering::*;
+
+        match *self {
+            OnlyTables(ref regexes) => !regexes.iter().any(|regex| regex.is_match(&name.sql_name)),
+            ExceptTables(ref regexes) => regexes.iter().any(|regex| regex.is_match(&name.sql_name)),
+            None => false,
+        }
+    }
+}
+
+impl<'de> Deserialize<'de> for Filtering {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct FilteringVisitor;
+
+        impl<'de> Visitor<'de> for FilteringVisitor {
+            type Value = Filtering;
+
+            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
+                f.write_str("either only_tables or except_tables")
+            }
+
+            fn visit_map<V>(self, mut map: V) -> Result<Self::Value, V::Error>
+            where
+                V: MapAccess<'de>,
+            {
+                let mut only_tables = None::<Vec<Regex>>;
+                let mut except_tables = None::<Vec<Regex>>;
+                while let Some(key) = map.next_key::<String>()? {
+                    match &key as &str {
+                        "only_tables" => {
+                            if only_tables.is_some() {
+                                return Err(de::Error::duplicate_field("only_tables"));
+                            }
+                            only_tables = Some(map.next_value()?);
+                        }
+                        "except_tables" => {
+                            if except_tables.is_some() {
+                                return Err(de::Error::duplicate_field("except_tables"));
+                            }
+                            except_tables = Some(map.next_value()?);
+                        }
+                        _ => {
+                            return Err(de::Error::unknown_field(
+                                &key,
+                                &["only_tables", "except_tables"],
+                            ))
+                        }
+                    }
+                }
+                match (only_tables, except_tables) {
+                    (Some(t), None) => Ok(Filtering::OnlyTables(t)),
+                    (None, Some(t)) => Ok(Filtering::ExceptTables(t)),
+                    (None, None) => Ok(Filtering::None),
+                    _ => Err(de::Error::duplicate_field("only_tables except_tables")),
+                }
+            }
+        }
+
+        deserializer.deserialize_map(FilteringVisitor)
     }
 }
