+++ diesel_cli/src/print_schema.rs
@@ -2,42 +2,13 @@ use crate::config;
 use crate::database::{Backend, InferConnection};
 use crate::infer_schema_internals::*;
 
-use serde::de::{self, MapAccess, Visitor};
-use serde::{Deserialize, Deserializer, Serialize};
-use serde_regex::Serde as RegexWrapper;
+use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::error::Error;
 use std::fmt::{self, Display, Formatter, Write};
 use std::io::Write as IoWrite;
 
 const SCHEMA_HEADER: &str = "// @generated automatically by Diesel CLI.\n";
-
-type Regex = RegexWrapper<::regex::Regex>;
-
-pub enum Filtering {
-    OnlyTables(Vec<Regex>),
-    ExceptTables(Vec<Regex>),
-    None,
-}
-
-#[allow(clippy::derivable_impls)] // that's not supported on rust 1.65
-impl Default for Filtering {
-    fn default() -> Self {
-        Filtering::None
-    }
-}
-
-impl Filtering {
-    pub fn should_ignore_table(&self, name: &TableName) -> bool {
-        use self::Filtering::*;
-
-        match *self {
-            OnlyTables(ref regexes) => !regexes.iter().any(|regex| regex.is_match(&name.sql_name)),
-            ExceptTables(ref regexes) => regexes.iter().any(|regex| regex.is_match(&name.sql_name)),
-            None => false,
-        }
-    }
-}
 
 /// How to sort columns when querying the table schema.
 #[derive(Debug, Deserialize, Serialize)]
@@ -769,61 +740,6 @@ impl<'a, 'b: 'a> Write for PadAdapter<'a, 'b> {
         }
 
         Ok(())
-    }
-}
-
-impl<'de> Deserialize<'de> for Filtering {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct FilteringVisitor;
-
-        impl<'de> Visitor<'de> for FilteringVisitor {
-            type Value = Filtering;
-
-            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                f.write_str("either only_tables or except_tables")
-            }
-
-            fn visit_map<V>(self, mut map: V) -> Result<Self::Value, V::Error>
-            where
-                V: MapAccess<'de>,
-            {
-                let mut only_tables = None::<Vec<Regex>>;
-                let mut except_tables = None::<Vec<Regex>>;
-                while let Some(key) = map.next_key::<String>()? {
-                    match &key as &str {
-                        "only_tables" => {
-                            if only_tables.is_some() {
-                                return Err(de::Error::duplicate_field("only_tables"));
-                            }
-                            only_tables = Some(map.next_value()?);
-                        }
-                        "except_tables" => {
-                            if except_tables.is_some() {
-                                return Err(de::Error::duplicate_field("except_tables"));
-                            }
-                            except_tables = Some(map.next_value()?);
-                        }
-                        _ => {
-                            return Err(de::Error::unknown_field(
-                                &key,
-                                &["only_tables", "except_tables"],
-                            ))
-                        }
-                    }
-                }
-                match (only_tables, except_tables) {
-                    (Some(t), None) => Ok(Filtering::OnlyTables(t)),
-                    (None, Some(t)) => Ok(Filtering::ExceptTables(t)),
-                    (None, None) => Ok(Filtering::None),
-                    _ => Err(de::Error::duplicate_field("only_tables except_tables")),
-                }
-            }
-        }
-
-        deserializer.deserialize_map(FilteringVisitor)
     }
 }
