+++ diesel_cli/tests/generate_migrations/diff_table_filter/diesel.toml
@@ -1,0 +1,2 @@
+[print_schema]
+filter = { only_tables = ["table_a"] }
