+++ diesel_cli/tests/migration_generate.rs
@@ -212,52 +212,52 @@ Creating custom_migrations.12345_stuff.down.sql
 
 #[test]
 fn migration_generate_from_diff_drop_table() {
-    test_generate_migration("diff_drop_table", Vec::new());
+    test_generate_migration("diff_drop_table", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_add_table() {
-    test_generate_migration("diff_add_table", Vec::new());
+    test_generate_migration("diff_add_table", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_drop_alter_table_add_column() {
-    test_generate_migration("diff_alter_table_add_column", Vec::new());
+    test_generate_migration("diff_alter_table_add_column", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_alter_table_drop_column() {
-    test_generate_migration("diff_alter_table_drop_column", Vec::new());
+    test_generate_migration("diff_alter_table_drop_column", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_add_table_with_fk() {
-    test_generate_migration("diff_add_table_with_fk", Vec::new());
+    test_generate_migration("diff_add_table_with_fk", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_drop_table_with_fk() {
-    test_generate_migration("diff_drop_table_with_fk", Vec::new());
+    test_generate_migration("diff_drop_table_with_fk", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_drop_table_all_the_types() {
-    test_generate_migration("diff_drop_table_all_the_types", Vec::new());
+    test_generate_migration("diff_drop_table_all_the_types", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_add_table_all_the_types() {
-    test_generate_migration("diff_add_table_all_the_types", Vec::new());
+    test_generate_migration("diff_add_table_all_the_types", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_add_table_composite_key() {
-    test_generate_migration("diff_add_table_composite_key", Vec::new());
+    test_generate_migration("diff_add_table_composite_key", Vec::new(), false);
 }
 
 #[test]
 fn migration_generate_from_diff_filter() {
-    test_generate_migration("diff_table_filter", Vec::new());
+    test_generate_migration("diff_table_filter", Vec::new(), true);
 }
 
 #[cfg(feature = "sqlite")]
@@ -276,8 +276,20 @@ fn backend_file_path(test_name: &str, file: &str) -> PathBuf {
         .join(file)
 }
 
-fn test_generate_migration(test_name: &str, args: Vec<&str>) {
-    let p = project(test_name).build();
+fn test_generate_migration(test_name: &str, args: Vec<&str>, with_config: bool) {
+    let p = match with_config {
+        true => {
+            let test_path = Path::new(env!("CARGO_MANIFEST_DIR"))
+                .join("tests")
+                .join("generate_migrations")
+                .join(test_name);
+            let config = read_file(&test_path.join("diesel.toml"));
+
+            project(test_name).file("diesel.toml", &config).build()
+        }
+        false => project(test_name).build(),
+    };
+
     let db = crate::support::database(&p.database_url());
 
     p.command("setup").run();
